Change terms acceptance mechanism
Prior to this patch, the user had to signify acceptance of the privacy policy before they could see the button which allowed them to submit the form.  The reason for doing it that way was that the guidance in the Service Design Manual says you shouldn't use disabled buttons because they are typically low-contrast, which means that they fail accessibility testing.

This patch removes the checkbox, and changes the wording on the button to signify that clicking it both accepts the data processing terms, and submits the form.  That's one less click for the user.